# Use block-based size for filecache entry size