--- a/server/remote_cache/digest/digest.go
+++ b/server/remote_cache/digest/digest.go
@@ -595,8 +595,9 @@ func Diff(s1 []*repb.Digest, s2 []*repb.Digest) (missingFromS1 []*repb.Digest, m
 }
 
 type randomDataMaker struct {
-	src rand.Source
-	val int64
+	src              rand.Source
+	compressionRatio float64
+	val              int64
 }
 
 func (r *randomDataMaker) Read(p []byte) (n int, err error) {
@@ -604,11 +605,10 @@ func (r *randomDataMaker) Read(p []byte) (n int, err error) {
 	offset := 0
 	for {
 		// Generate a new random int64 (8 bytes) if we haven't generated one
-		// yet, or with a percent chance given by `percentCompressibility`. This
-		// is a *very* rough way to generate blobs with the average compression
+		// yet, or with a percent chance given by the compression ratio. This is
+		// a *very* rough way to generate blobs with the average compression
 		// ratios that we see in practice.
-		const percentCompressibility = 70
-		if r.val == 0 || r.src.Int63()%100 > percentCompressibility {
+		if r.val == 0 || r.src.Int63()%100 >= int64(r.compressionRatio*100) {
 			r.val = int64(r.src.Int63())
 		}
 		val := r.val
@@ -630,10 +630,25 @@ type Generator struct {
 }
 
 // RandomGenerator returns a digest sample generator for use in testing tools.
+// It generates digests with compression ratios similar to what we see in
+// practice.
 func RandomGenerator(seed int64) *Generator {
 	return &Generator{
-		randMaker: &randomDataMaker{src: rand.NewSource(seed)},
-		mu:        sync.Mutex{},
+		randMaker: &randomDataMaker{
+			src:              rand.NewSource(seed),
+			compressionRatio: 0.7,
+		},
+	}
+}
+
+// UniformRandomGenerator generates uniformly random, incompressible digests,
+// for use in testing. The data generated does not look realistic, but it is
+// useful in cases where unique digests are needed.
+func UniformRandomGenerator(seed int64) *Generator {
+	return &Generator{
+		randMaker: &randomDataMaker{
+			src: rand.NewSource(seed),
+		},
 	}
 }