--- a/server/util/disk/disk_windows.go
+++ b/server/util/disk/disk_windows.go
@@ -4,7 +4,9 @@ package disk
 
 import (
 	"fmt"
+	"os"
 
+	"github.com/buildbuddy-io/buildbuddy/server/util/status"
 	"golang.org/x/sys/windows"
 )
 
@@ -24,4 +26,14 @@ func GetDirUsage(path string) (*DirUsage, error) {
 		FreeBytes:  totalNumberOfFreeBytes,
 		AvailBytes: freeBytesAvailableToCaller,
 	}, nil
+}
+
+// EstimatedFileDiskUsage returns an estimate of the disk usage required for
+// the given regular file info.
+func EstimatedFileDiskUsage(info os.FileInfo) (int64, error) {
+	if !info.Mode().IsRegular() {
+		return 0, status.InvalidArgumentError("not a regular file")
+	}
+	// TODO: figure out something better for Windows.
+	return info.Size(), nil
 }