When calling `AddFile`, stat the file to get the number of blocks consumed by the file, and compute the size based on the number of blocks consumed, rather than the content stored in the blocks. The extra `stat()` syscall does not significantly affect performance according to benchmarks:

```
goos: linux
goarch: amd64
cpu: AMD Ryzen 9 7950X 16-Core Processor            
                                   │ /tmp/bench_false.txt │        /tmp/bench_true.txt         │
                                   │        sec/op        │   sec/op     vs base               │
FilecacheLink/100%Link/0%Add/1K-32            15.52m ± 1%   15.57m ± 1%       ~ (p=0.631 n=10)
FilecacheLink/95%Link/5%Add/1K-32             14.77m ± 2%   14.75m ± 3%       ~ (p=0.436 n=10)
geomean                                       15.14m        15.15m       +0.07%
```

A future improvement could be to also account for the file metadata, not just the file contents. But just computing the block-based size should be good enough for now.

**Related issues**:
* Fixes https://github.com/buildbuddy-io/buildbuddy-internal/issues/3300
* Fixes https://github.com/buildbuddy-io/buildbuddy-internal/issues/3248